Don't claim rudimentary robustness in SimplePedPop
I think this is not really a feature in the sense that it's only
against very stupid misbehaving signers. I see it more of a debug
feature, and a necessary consequence of the fact that we need to
abort in this case, and we happen to know who has sent the invalid
signature.