md: Kill footnote on Repairable Threshold Schemes
What we have built is, in fact, a repairable threshold scheme, where
other participants can help repair a share of a broken participant.
(Our scheme is even better because a single peer suffices.) The footnote
was a bit misleading in that it says that this could be possible without
the seed. But without some secret, the participant can't authenticate to
peers anymore, no matter what repairable threshold scheme will be used.

(If we changed the scheme to store also the "self-share" in encrypted
form in the recovery data, then we'd obtain a scheme where the seed is
really only used for "authentication", namely decrypting the secshare.)