secp256k1proto: deserialize x-only-pk with `from_bytes_xonly` (schnorr_verify)
If the passed x-only-pubkey overflows (x >= p), this method fails
immediately rather than silently wrapping around and then likely
failing later (either because x % p is not on the curve or because the
actual signature verification fails). For the user of `schnorr_verify`
this shouldn't make any difference, as `False` is returned with both
variants, so this is probably more of a cosmetic change.